+++ Final/Backend/QuanLyKTX/src/main/java/web/nhom8/quanlyktx/model/RoomPaymentModel.java
@@ -3,8 +3,8 @@ package web.nhom8.quanlyktx.model;
 import java.sql.Date;
 
 public class RoomPaymentModel {
-    private int PaymentId;
-    private int StudentRoomId;
+    private Long PaymentId;
+    private Long StudentRoomId;
     private float PaymentMoney;
     private Date PaymentDate;
     private int State;
@@ -12,7 +12,7 @@ public class RoomPaymentModel {
     public RoomPaymentModel() {
 
     }
-    public RoomPaymentModel(int paymentId, int studentRoomId, float paymentMoney, Date paymentDate, int state) {
+    public RoomPaymentModel(Long paymentId, Long studentRoomId, float paymentMoney, Date paymentDate, int state) {
         PaymentId = paymentId;
         StudentRoomId = studentRoomId;
         PaymentMoney = paymentMoney;
@@ -20,19 +20,19 @@ public class RoomPaymentModel {
         State = state;
     }
 
-    public int getPaymentId() {
+    public Long getPaymentId() {
         return PaymentId;
     }
 
-    public void setPaymentId(int paymentId) {
+    public void setPaymentId(Long paymentId) {
         PaymentId = paymentId;
     }
 
-    public int getStudentRoomId() {
+    public Long getStudentRoomId() {
         return StudentRoomId;
     }
 
-    public void setStudentRoomId(int studentRoomId) {
+    public void setStudentRoomId(Long studentRoomId) {
         StudentRoomId = studentRoomId;
     }
