+++ Final/Backend/QuanLyKTX/src/main/java/web/nhom8/quanlyktx/model/UtilityBillModel.java
@@ -3,8 +3,8 @@ package web.nhom8.quanlyktx.model;
 import java.sql.Date;
 
 public class UtilityBillModel {
-    private int BillId;
-    private int RoomId;
+    private Long BillId;
+    private Long RoomId;
     private float Money;
     private Date PaymentDate;
     private int State;
@@ -12,7 +12,7 @@ public class UtilityBillModel {
     public UtilityBillModel() {
 
     }
-    public UtilityBillModel(int billId, int roomId, float money, Date paymentDate, int state) {
+    public UtilityBillModel(Long billId, Long roomId, float money, Date paymentDate, int state) {
         BillId = billId;
         RoomId = roomId;
         Money = money;
@@ -20,18 +20,18 @@ public class UtilityBillModel {
         State = state;
     }
 
-    public int getBillId() {
+    public Long getBillId() {
         return BillId;
     }
 
-    public void setBillId(int billId) {
+    public void setBillId(Long billId) {
         BillId = billId;
     }
-    public int getRoomId() {
+    public Long getRoomId() {
         return RoomId;
     }
 
-    public void setRoomId(int roomId) {
+    public void setRoomId(Long roomId) {
         RoomId = roomId;
     }
